requirements.txt
fastapi
uvicorn
python-dotenv
requests
Jinja2